Change food price from Integer to  Double

+++ src/test/java/restful/api/ezdine/controller/FoodControllerTest.java
@@ -77,7 +77,7 @@ public class FoodControllerTest {
 
     private final String foodCode = "spaghetti";
     private final String foodName = "spaghetti bolognese";
-    private final Integer foodPrice = 20;
+    private final Double foodPrice = 20.0;
     private final Integer foodStock = 0;
 
     @BeforeEach
